# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Locket Clone API Contract
  description: "Tài liệu đặc tả API chuẩn OpenAPI 3.0 / Swagger UI dành cho Frontend\
    \ (Web & Mobile). Cung cấp đầy đủ các endpoint xác thực (JWT), chia sẻ ảnh (Locket),\
    \ quản lý giao dịch tài chính, kết bạn và lưu trữ ảnh Garage S3."
  contact:
    name: Development Team
    email: dev@locketclone.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development Server
security:
- Bearer Authentication: []
...
paths:
  /api/v1/photos/{photoId}/reactions/me:
    put:
      tags:
      - Photos
      summary: Thả cảm xúc / reaction vào ảnh
      description: "Thêm hoặc đổi reaction (LIKE, LOVE, HAHA, WOW, SAD, ANGRY)."
      operationId: upsertMyReaction
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPhotoReactionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoReactionResponse"
    delete:
      tags:
      - Photos
      summary: Xóa cảm xúc đã thả
      description: Gỡ bỏ reaction của người dùng hiện tại khỏi bức ảnh.
      operationId: removeMyReaction
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
  /api/v1/photos:
    post:
      tags:
      - Photos
      summary: Đăng tải ảnh khoảnh khắc mới
      description: "Upload ảnh (JPEG/PNG/WEBP) kèm thông tin tài chính (số tiền, danh\
        \ mục, loại giao dịch, ghi chú) và phạm vi bạn bè nhận ảnh (ALL_FRIENDS hoặ\
        c SELECTED_FRIENDS)."
      operationId: uploadPhoto
      parameters:
      - name: caption
        in: query
        required: false
        schema:
          type: string
      - name: amount
        in: query
        required: false
        schema:
          type: number
      - name: transactionType
        in: query
        required: false
        schema:
          type: string
      - name: note
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: recipientScope
        in: query
        required: false
        schema:
          type: string
          enum:
          - ALL_FRIENDS
          - SELECTED_FRIENDS
      - name: audienceMode
        in: query
        required: false
        schema:
          type: string
          enum:
          - ALL_FRIENDS
          - SELECTED_FRIENDS
      - name: recipientIds
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: takenAt
        in: query
        required: false
        schema:
          type: string
          format: date-time
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoResponse"
  /api/v1/friend-invite-links:
    post:
      tags:
      - Friend Invites
      summary: Tạo hoặc làm mới link mời kết bạn
      description: "Sinh link mời kèm token ngẫu nhiên, cấu hình số lượt sử dụng tố\
        i đa và thời gian hết hạn (TTL)."
      operationId: createOrRotate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFriendInviteLinkRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FriendInviteLinkResponse"
  /api/v1/friend-invite-links/accept:
    post:
      tags:
      - Friend Invites
      summary: Chấp nhận kết bạn qua link
      description: Chấp nhận kết bạn bằng token trong link mời (có chống race condition
        bằng pessimistic lock).
      operationId: acceptByLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptFriendInviteLinkRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AcceptFriendInviteLinkResponse"
...
  /api/v1/photos/{photoId}/transaction:
    patch:
      tags:
      - Photos
      summary: Cập nhật toàn bộ thông tin giao dịch ảnh
      description: "Chỉnh sửa caption, số tiền, loại giao dịch, danh mục, thời gian\
        \ của ảnh."
      operationId: updatePhotoTransaction
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoTransactionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoResponse"
  /api/v1/photos/{photoId}/expense:
    patch:
      tags:
      - Photos
      summary: Cập nhật chi phí gắn với ảnh
      description: "Chỉnh sửa số tiền, ghi chú, danh mục tài chính của bức ảnh."
      operationId: updatePhotoExpense
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoExpenseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoResponse"
...
  /api/v1/photos/{photoId}:
    get:
      tags:
      - Photos
      summary: Xem chi tiết một ảnh
      description: "Lấy thông tin chi tiết một bức ảnh bao gồm người gửi, danh mụ\
        c, số tiền, ghi chú."
      operationId: getPhotoDetail
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoResponse"
    delete:
      tags:
      - Photos
      summary: Xóa ảnh / giao dịch
      description: Đánh dấu xóa ảnh (DELETED) và tự động xóa file lưu trữ trên Garage
        S3.
      operationId: deleteTransaction
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/photos/{photoId}/reactions/summary:
    get:
      tags:
      - Photos
      summary: Thống kê cảm xúc của ảnh
      description: Tổng hợp số lượng reaction theo từng loại và danh sách người thả
        cảm xúc.
      operationId: getReactionSummary
      parameters:
      - name: photoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PhotoReactionSummaryResponse"
  /api/v1/photos/my-photos:
    get:
      tags:
      - Photos
      summary: Lấy danh sách ảnh của tôi
      description: Phân trang danh sách ảnh mà người dùng hiện tại đã đăng tải.
      operationId: getMyPhotos
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagePhotoResponse"
  /api/v1/photos/me:
    get:
      tags:
      - Photos
      summary: Lấy danh sách ảnh của tôi
      description: Phân trang danh sách ảnh mà người dùng hiện tại đã đăng tải.
      operationId: getMyPhotos_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagePhotoResponse"
  /api/v1/photos/feed:
    get:
      tags:
      - Photos
      summary: Lấy feed ảnh bạn bè
      description: Cuộn vô tận (Slice) danh sách ảnh nhận được từ bạn bè hoặc lọc
        theo từng bạn cụ thể.
      operationId: getFeedPhotos
      parameters:
      - name: friendId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SlicePhotoResponse"
...
components:
  schemas:
    UpsertPhotoReactionRequest:
      type: object
      properties:
        type:
          type: string
      required:
      - type
    PhotoReactionResponse:
      type: object
      properties:
        photoId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
        createdAt:
          type: string
          format: date-time
...
    PhotoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        senderDisplayName:
          type: string
        senderAvatarUrl:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        caption:
          type: string
        note:
          type: string
        amount:
          type: number
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        recipientScope:
          type: string
          enum:
          - ALL_FRIENDS
          - SELECTED_FRIENDS
        recipientCount:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - READY
          - FAILED
          - DELETED
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        takenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateFriendInviteLinkRequest:
      type: object
      properties:
        ttlMinutes:
          type: integer
          format: int32
          maximum: 43200
          minimum: 10
    FriendInviteLinkResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inviteUrl:
          type: string
        tokenPreview:
          type: string
        maxUses:
          type: integer
          format: int32
        usedCount:
          type: integer
          format: int32
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
        status:
          type: string
    AcceptFriendInviteLinkRequest:
      type: object
      properties:
        token:
          type: string
      required:
      - token
    AcceptFriendInviteLinkResponse:
      type: object
      properties:
        friendshipId:
          type: string
          format: uuid
        status:
          type: string
        friend:
          $ref: "#/components/schemas/UserResponse"
        acceptedAt:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        isVerified:
          type: boolean
        isGoldMember:
          type: boolean
        profileCompleted:
          type: boolean
...
    UpdatePhotoTransactionRequest:
      type: object
      properties:
        caption:
          type: string
          maxLength: 1000
          minLength: 0
        amount:
          type: number
        note:
          type: string
          maxLength: 255
          minLength: 0
        categoryId:
          type: string
          format: uuid
        clearCategory:
          type: boolean
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
        takenAt:
          type: string
          format: date-time
        occurredAt:
          type: string
          format: date-time
    UpdatePhotoExpenseRequest:
      type: object
      properties:
        amount:
          type: number
        note:
          type: string
          maxLength: 255
          minLength: 0
        categoryId:
          type: string
          format: uuid
...
    PhotoReactionSummaryResponse:
      type: object
      properties:
        photoId:
          type: string
          format: uuid
        totalCount:
          type: integer
          format: int64
        myReaction:
          type: string
        countsByType:
          type: object
          additionalProperties:
            type: integer
            format: int64
        reactors:
          type: array
          items:
            $ref: "#/components/schemas/PhotoReactorResponse"
    PhotoReactorResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        avatarUrl:
          type: string
        type:
          type: string
        reactedAt:
          type: string
          format: date-time
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PagePhotoResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/PhotoResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    SlicePhotoResponse:
      type: object
      properties:
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/PhotoResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Nhập Access Token (JWT) được trả về từ API /api/v1/auth/login hoặc
        /api/v1/auth/verify
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT